requirements.txt
pyramid
waitress
google-cloud-ndb
google-cloud-logging
